+++ RichardEngine/include/richard.h
@@ -2,4 +2,4 @@
 * This header contains all the headers that will be included in RichardEditor
 */
 
-#include "engine.h"
+#include "core/engine.h"
