+++ RichardEngine/src/core/constants.h
