+++ RichardEngine/src/core/engine.cpp
@@ -1,5 +1,5 @@
 #include "engine.h"
-#include "managers/constants.h"
+#include "constants.h"
 #include "SDL.h"
 #include <iostream>
 #include <string>
