+++ RichardEngine/src/core/window.cpp
